requirements.txt
torch
torchvision
transformers
pillow
azure-storage-blob
azure-storage-queue
numpy